+++ js/main.js
@@ -6,7 +6,7 @@ window.CQ = window.CQ || {};
 
   CQ.state = {
     game: null, screen: "cover",
-    ctab: "attrs", ttab: "liga", clubTab: "info", feedFilter: "all",
+    ctab: "attrs", ttab: "liga", clubTab: "info", feedFilter: "all", calFilter: "all",
     draft: null, live: null, itv: null, lifeEv: null, champFilter: "BRA", summary: null,
     lastSeen: { money: null, fame: null, marcos: {} }
   };